First Commit
This commit involves the use of autoencoders for anomaly detection for credit card fraud detection. Unfortunately the important metric which is precision was very low in this case. There are improvements still being made to this. One main reason I think the precision was low was due to data imbalances and I plan to employ the use of SMOTE to counter this issue.